requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
pydantic
motor
psycopg2
logger
alembic
python-dotenv
pyyaml